Move SpectralSources et al. to a submodule.
