+++ sbpy/spectroscopy/__init__.py
@@ -3,5 +3,6 @@ SBPy Module for small body spectroscopy
 """
 
 from .core import *
+from .sources import *
 from .sun import *
 from .vega import *
